pyproject.toml
[tool.poetry]
name = "biblioteca"
version = "0.1.0"
description = ""
authors = ["Alexander Alcázar <aalcazarprogramacion@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = ">=4.2,<5.0"
python-dotenv = "^1.2.1"
psycopg2-binary = "^2.9.11"
djangorestframework = "^3.16.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
